app/requirements.txt
openai==1.55.3
python-telegram-bot==22.2
langsmith==0.4.8
langgraph==0.5.3
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langgraph-checkpoint-postgres==2.0.23
psycopg==3.2.9
psycopg-pool==3.2.6